% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_LG_api.R
\name{call_LG_api}
\alias{call_LG_api}
\title{Call the LibGuides GET guides API (v1.1).}
\usage{
call_LG_api(siteID, key, content, options = "")
}
\arguments{
\item{siteID}{numeric or character string containing your institution's site ID}

\item{key}{character string containing your institution's API key}

\item{content}{character string: one of "guides", "assets", "accounts", or "subjects"}

\item{options}{character string containing additional API options (must start with &)}
}
\value{
Character string containing the JSON response from the LibGuides API.
}
\description{
This function calls the LibGuides GET guides API, and returns a character
string containing the JSON. This can then be written to a text file, or
passed to other methods.
}
\details{
The LibGuides API requires a LibGuides CMS subscription, so if your
institution does not have a CMS subscription, this function will not work.
Your site ID and API key can be found under Tools > API; click the GET Guides link.

## Guides
Requests for guides will automatically include guide pages and guide owners.
}
\examples{
output <- call_LG_api(siteID="YOUR_SITE_ID", key="YOUR_API_KEY", content="guides")
}
