% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_database_assets.R
\name{get_database_assets}
\alias{get_database_assets}
\title{Call the LibGuides GET assets API (v1.1) and simplify it.}
\usage{
get_database_assets(siteID = NULL, key = NULL, options = "")
}
\arguments{
\item{siteID}{numeric or character string containing your institution's site ID}

\item{key}{character string containing your institution's API key}

\item{options}{character string - any other options to append to the API request (see the API documentation for all options for each GET routine.)}
}
\value{
Data frame containing the book asset data. If desired, can be joined with the user account data on the variable \code{owner_id}.
}
\description{
This function calls the LibGuides GET assets API using a filter for database assets, and returns a dataframe of the matching assets.
}
\details{
The LibGuides API requires a LibGuides CMS subscription, so if your institution does not have a CMS subscription, this function will not work. Your site ID and API key can be found under Tools > API; click the GET Guides link.
}
\examples{
mydata <- get_database_assets(siteID="YOUR_SITEID_HERE", key="YOUR_KEY_HERE")
}
