% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_guides_data.R
\name{get_guides_data}
\alias{get_guides_data}
\title{Call the LibGuides GET guides API (v1.1) and simplify it.}
\usage{
get_guides_data(siteID = NULL, key = NULL, expand_pages = TRUE)
}
\arguments{
\item{siteID}{numeric or character string containing your institution's site ID}

\item{key}{character string containing your institution's API key}

\item{expand_pages}{logical - if TRUE, output dataframe will have one row per page per guide; if FALSE, output dataframe will have one row per guide. Default is TRUE.}
}
\value{
Data frame containing the guides data (at either the guide level or page level), with the "owner" and "group" columns expanded.
}
\description{
This function calls the LibGuides GET guides API, and returns a dataframe with your institution's guide data.
}
\details{
The LibGuides API requires a LibGuides CMS subscription, so if your institution does not have a CMS subscription, this function will not work. Your site ID and API key can be found under Tools > API; click the GET Guides link.
}
\examples{
mydata <- get_guides_data(siteID="YOUR_SITEID_HERE", key="YOUR_KEY_HERE")
}
